requirements.txt
boto3==1.26.150
bs4==0.0.1
faiss-cpu==1.7.4
google-api-python-client==2.103.0
openai==0.27.8
python-dotenv==1.0.0
tiktoken==0.4.0
tabulate==0.9.0
matplotlib==3.7.1
streamlit==1.23.1
slack_bolt==1.18.0
langchain==0.0.228
